+++ app/assets/javascripts/admin/admin_rest_api.js
@@ -25,8 +25,8 @@ import type {
   APIOpenTasksReportType,
   APITracingType,
 } from "admin/api_flow_types";
-import type { QueryObjectType } from "admin/views/task/task_search_form";
-import type { NewTaskType, TaskCreationResponseType } from "admin/views/task/task_create_bulk_view";
+import type { QueryObjectType } from "admin/task/task_search_form";
+import type { NewTaskType, TaskCreationResponseType } from "admin/task/task_create_bulk_view";
 
 const MAX_SERVER_ITEMS_PER_RESPONSE = 1000;
