+++ app/assets/javascripts/admin/auth/auth_token_view.js
