+++ app/assets/javascripts/admin/auth/change_password_view.js
