+++ app/assets/javascripts/admin/auth/finish_reset_password_view.js
