+++ app/assets/javascripts/admin/auth/login_view.js
