+++ app/assets/javascripts/admin/auth/registration_view.js
