+++ app/assets/javascripts/admin/auth/start_reset_password_view.js
