+++ app/assets/javascripts/admin/dataset/dataset_upload_view.js
