+++ app/assets/javascripts/admin/help/keyboardshortcut_view.js
