+++ app/assets/javascripts/admin/project/project_create_view.js
