+++ app/assets/javascripts/admin/project/project_list_view.js
