+++ app/assets/javascripts/admin/scripts/script_create_view.js
