+++ app/assets/javascripts/admin/scripts/script_list_view.js
