+++ app/assets/javascripts/admin/statistic/open_tasks_report_view.js
