+++ app/assets/javascripts/admin/statistic/project_progress_report_view.js
