+++ app/assets/javascripts/admin/statistic/statistic_view.js
