+++ app/assets/javascripts/admin/statistic/team_selection_form.js
