+++ app/assets/javascripts/admin/task/task_create_view.js
@@ -1,8 +1,8 @@
 // @flow
 import React from "react";
 import { Tabs, Icon } from "antd";
-import TaskCreateFormView from "admin/views/task/task_create_form_view";
-import TaskCreateBulkView from "admin/views/task/task_create_bulk_view";
+import TaskCreateFormView from "admin/task/task_create_form_view";
+import TaskCreateBulkView from "admin/task/task_create_bulk_view";
 
 const { TabPane } = Tabs;
