+++ app/assets/javascripts/admin/tasktype/task_type_create_view.js
