+++ app/assets/javascripts/admin/tasktype/task_type_list_view.js
