+++ app/assets/javascripts/admin/team/create_team_modal_view.js
