+++ app/assets/javascripts/admin/time/time_line_view.js
