+++ app/assets/javascripts/admin/user/user_list_view.js
@@ -5,8 +5,8 @@ import _ from "lodash";
 import * as React from "react";
 import { Link, withRouter } from "react-router-dom";
 import { Table, Tag, Icon, Spin, Button, Input } from "antd";
-import TeamRoleModalView from "admin/views/user/team_role_modal_view";
-import ExperienceModalView from "admin/views/user/experience_modal_view";
+import TeamRoleModalView from "admin/user/team_role_modal_view";
+import ExperienceModalView from "admin/user/experience_modal_view";
 import TemplateHelpers from "libs/template_helpers";
 import Utils from "libs/utils";
 import { getEditableUsers, updateUser } from "admin/admin_rest_api";
