+++ app/assets/javascripts/dashboard/advanced_dataset/dataset_access_list_view.js
