+++ app/assets/javascripts/dashboard/advanced_dataset/dataset_action_view.js
