+++ app/assets/javascripts/dashboard/dataset/dataset_import_view.js
