+++ app/assets/javascripts/dashboard/dataset/team_assignment_modal.js
@@ -4,7 +4,7 @@ import * as React from "react";
 import { Modal, Spin, Select } from "antd";
 import Request from "libs/request";
 import type { APITeamType } from "admin/api_flow_types";
-import type { DatasetType } from "dashboard/views/dataset_view";
+import type { DatasetType } from "dashboard/dataset_view";
 
 const { Option } = Select;
