+++ app/assets/javascripts/dashboard/dataset_view.js
@@ -7,8 +7,8 @@ import Request from "libs/request";
 import Utils from "libs/utils";
 import moment from "moment";
 import { Spin, Input, Button } from "antd";
-import AdvancedDatasetView from "dashboard/views/advanced_dataset/advanced_dataset_view";
-import GalleryDatasetView from "dashboard/views/gallery_dataset_view";
+import AdvancedDatasetView from "dashboard/advanced_dataset/advanced_dataset_view";
+import GalleryDatasetView from "dashboard/gallery_dataset_view";
 import Persistence from "libs/persistence";
 import { PropTypes } from "prop-types";
 import type { APIUserType, APIDatasetType } from "admin/api_flow_types";
