+++ app/assets/javascripts/dashboard/explorative_annotations_view.js
