+++ app/assets/javascripts/dashboard/spotlight_view.js
@@ -2,10 +2,10 @@
 import * as React from "react";
 import { Link } from "react-router-dom";
 import { Spin, Layout, Row, Col } from "antd";
-import { transformDatasets } from "dashboard/views/dataset_view";
-import GalleryDatasetView from "dashboard/views/gallery_dataset_view";
+import { transformDatasets } from "dashboard/dataset_view";
+import GalleryDatasetView from "dashboard/gallery_dataset_view";
 import Request from "libs/request";
-import type { DatasetType } from "dashboard/views/dataset_view";
+import type { DatasetType } from "dashboard/dataset_view";
 
 const { Header, Content, Footer } = Layout;
