+++ app/assets/javascripts/dashboard/transfer_task_modal.js
