+++ app/assets/javascripts/router.js
@@ -15,34 +15,34 @@ import Navbar from "navbar";
 import Imprint from "components/imprint";
 
 import TracingLayoutView from "oxalis/view/tracing_layout_view";
-import DashboardView from "dashboard/views/dashboard_view";
-import SpotlightView from "dashboard/views/spotlight_view";
-import LoginView from "admin/views/auth/login_view";
-import RegistrationView from "admin/views/auth/registration_view";
-import StartResetPasswordView from "admin/views/auth/start_reset_password_view";
-import FinishResetPasswordView from "admin/views/auth/finish_reset_password_view";
-import ChangePasswordView from "admin/views/auth/change_password_view";
-import AuthTokenView from "admin/views/auth/auth_token_view";
-import DatasetImportView from "dashboard/views/dataset/dataset_import_view";
+import DashboardView from "dashboard/dashboard_view";
+import SpotlightView from "dashboard/spotlight_view";
+import LoginView from "admin/auth/login_view";
+import RegistrationView from "admin/auth/registration_view";
+import StartResetPasswordView from "admin/auth/start_reset_password_view";
+import FinishResetPasswordView from "admin/auth/finish_reset_password_view";
+import ChangePasswordView from "admin/auth/change_password_view";
+import AuthTokenView from "admin/auth/auth_token_view";
+import DatasetImportView from "dashboard/dataset/dataset_import_view";
 
 // admin
-import KeyboardShortcutView from "admin/views/help/keyboardshortcut_view";
-import DatasetAddView from "admin/views/dataset/dataset_add_view";
-import UserListView from "admin/views/user/user_list_view";
-import TeamListView from "admin/views/team/team_list_view";
-import TaskListView from "admin/views/task/task_list_view";
-import TaskTypeListView from "admin/views/tasktype/task_type_list_view";
-import ProjectListView from "admin/views/project/project_list_view";
-import StatisticView from "admin/views/statistic/statistic_view";
-import ProjectProgressReportView from "admin/views/statistic/project_progress_report_view";
-import OpenTasksReportView from "admin/views/statistic/open_tasks_report_view";
-import ScriptListView from "admin/views/scripts/script_list_view";
-import ProjectCreateView from "admin/views/project/project_create_view";
-import TaskCreateView from "admin/views/task/task_create_view";
-import TaskCreateFormView from "admin/views/task/task_create_form_view";
-import TaskTypeCreateView from "admin/views/tasktype/task_type_create_view";
-import ScriptCreateView from "admin/views/scripts/script_create_view";
-import TimeLineView from "admin/views/time/time_line_view";
+import KeyboardShortcutView from "admin/help/keyboardshortcut_view";
+import DatasetAddView from "admin/dataset/dataset_add_view";
+import UserListView from "admin/user/user_list_view";
+import TeamListView from "admin/team/team_list_view";
+import TaskListView from "admin/task/task_list_view";
+import TaskTypeListView from "admin/tasktype/task_type_list_view";
+import ProjectListView from "admin/project/project_list_view";
+import StatisticView from "admin/statistic/statistic_view";
+import ProjectProgressReportView from "admin/statistic/project_progress_report_view";
+import OpenTasksReportView from "admin/statistic/open_tasks_report_view";
+import ScriptListView from "admin/scripts/script_list_view";
+import ProjectCreateView from "admin/project/project_create_view";
+import TaskCreateView from "admin/task/task_create_view";
+import TaskCreateFormView from "admin/task/task_create_form_view";
+import TaskTypeCreateView from "admin/tasktype/task_type_create_view";
+import ScriptCreateView from "admin/scripts/script_create_view";
+import TimeLineView from "admin/time/time_line_view";
 
 import type { OxalisState } from "oxalis/store";
 import type { APITracingType, APIUserType } from "admin/api_flow_types";
